+++ priv/repo/migrations/20200902113604_add_approval_note_to_message_approvals_projection.exs
@@ -1,0 +1,9 @@
+defmodule DealogBackoffice.Repo.Migrations.AddApprovalNoteToMessageApprovalsProjection do
+  use Ecto.Migration
+
+  def change do
+    alter table(:message_approvals) do
+      add :note, :text
+    end
+  end
+end